requirements/base.txt
Django==5.2.5
djangorestframework==3.16.1
python-decouple==3.8
djangorestframework_simplejwt==5.5.1